server/requirements.txt
pydantic~=2.11.7
pydantic-settings~=2.10.1
fastapi~=0.115.13
httpx~=0.28.1
numpy~=2.3.0
uvicorn
sentence-transformers~=4.1.0
slowapi~=0.1.9
google-genai